Typehint and phpdoc improvements based on phpstan
Fix low hanging fruit from running phpstan with level 2 checking.
Also adds a basic psalm config and a new composer run-script target
('analyse') to execute both tools.

Tool dependencies are intentially not added to avoid unnecessary bloat
and download time in travis.

This is intended as a purely manual tool for regression checking and a
help to iteratively improve the codebase.