Removed error catching on invalid attributes
The error catching made it hard to find the reason for "ignored" attributes. This change addresses the issues #1267 and #1162.

Instead of catching errors of non existing attribute classes, the existence of the class is checked before the attribute is tried to instantiate. This reveals invalid attribute definitions, but there won't be shown errors for attributes whose class definitions do not exist.